refactor: 분리된 advice 통일
