--- src/main/java/roomescape/presentation/advice/GlobalExceptionHandler.java
@@ -1,23 +1,0 @@
-package roomescape.presentation.advice;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ProblemDetail;
-import org.springframework.web.ErrorResponse;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-    private static final Logger logger = LoggerFactory.getLogger(GlobalExceptionHandler.class);
-
-    @ExceptionHandler(Exception.class)
-    public ProblemDetail handleException(Exception exception) {
-        logger.error(exception.getMessage(), exception);
-        if (exception instanceof ErrorResponse errorResponse) {
-            return ProblemDetail.forStatusAndDetail(errorResponse.getStatusCode(), exception.getMessage());
-        }
-        return ProblemDetail.forStatusAndDetail(HttpStatus.INTERNAL_SERVER_ERROR, "예기치 않은 오류가 발생했습니다.");
-    }
-}
